requirements.txt
customtkinter==5.2.1
pynput==1.7.6
pystray==0.19.5
pyperclip==1.8.2
Pillow==10.3.0
pytest==7.4.3
